+++ src/main/java/dev/willyelton/crystal_tools/levelable/CrystalBackpack.java
@@ -40,7 +40,6 @@ import net.minecraftforge.items.ItemHandlerHelper;
 import net.minecraftforge.items.ItemStackHandler;
 import net.minecraftforge.network.NetworkHooks;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -126,7 +125,7 @@ public class CrystalBackpack extends Item implements LevelableItem {
     }
 
     @Override
-    public @Nullable ICapabilityProvider initCapabilities(ItemStack stack, @Nullable CompoundTag nbt) {
+    public ICapabilityProvider initCapabilities(ItemStack stack, CompoundTag nbt) {
         return new CrystalBackpackCapabilityProvider(stack);
     }