+++ src/main/java/dev/willyelton/crystal_tools/network/packet/BlockAttributePacket.java
@@ -14,11 +14,13 @@ public class BlockAttributePacket {
     private final String key;
     private final float value;
     private final int id;
+    private final int pointsToSpend;
 
-    public BlockAttributePacket(String key, float value, int id) {
+    public BlockAttributePacket(String key, float value, int id, int pointsToSpend) {
         this.key = key;
         this.value = value;
         this.id = id;
+        this.pointsToSpend = pointsToSpend;
     }
 
     public BlockAttributePacket(FriendlyByteBuf buffer) {
@@ -26,6 +28,7 @@ public class BlockAttributePacket {
         this.key = buffer.readCharSequence(keyLen, Charset.defaultCharset()).toString();
         this.value = buffer.readFloat();
         this.id = buffer.readInt();
+        this.pointsToSpend = buffer.readInt();
     }
 
     public void encode(FriendlyByteBuf buffer) {
@@ -33,6 +36,7 @@ public class BlockAttributePacket {
         buffer.writeCharSequence(this.key, Charset.defaultCharset());
         buffer.writeFloat(this.value);
         buffer.writeInt(this.id);
+        buffer.writeInt(this.pointsToSpend);
     }
 
     public void handle(Supplier<NetworkEvent.Context> ctx) {
@@ -43,10 +47,12 @@ public class BlockAttributePacket {
 
             if (container instanceof CrystalFurnaceContainerMenu crystalFurnaceContainerMenu) {
                 CrystalFurnaceBlockEntity blockEntity = crystalFurnaceContainerMenu.getBlockEntity();
+                int skillPoints = blockEntity.getSkillPoints();
+                int pointsToAdd = Math.min(skillPoints, this.pointsToSpend);
 
-                blockEntity.addToData(this.key, this.value);
+                blockEntity.addToData(this.key, this.value * pointsToAdd);
                 if (this.id != -1) {
-                    blockEntity.addToPoints(this.id, 1);
+                    blockEntity.addToPoints(this.id, pointsToAdd);
                 }
             }
         }