+++ src/main/java/dev/willyelton/crystal_tools/network/packet/ToolAttributePacket.java
@@ -21,16 +21,14 @@ public class ToolAttributePacket {
     private final float value;
     private final int id;
     private final int slotIndex;
+    private final int pointsToSpend;
 
-    public ToolAttributePacket(String key, float value, int id) {
-        this(key, value, id, -1);
-    }
-
-    public ToolAttributePacket(String key, float value, int id, int slotIndex) {
+    public ToolAttributePacket(String key, float value, int id, int slotIndex, int pointsToSpend) {
         this.key = key;
         this.value = value;
         this.id = id;
         this.slotIndex = slotIndex;
+        this.pointsToSpend = pointsToSpend;
     }
 
     public ToolAttributePacket(FriendlyByteBuf buffer) {
@@ -39,6 +37,7 @@ public class ToolAttributePacket {
         this.value = buffer.readFloat();
         this.id = buffer.readInt();
         this.slotIndex = buffer.readInt();
+        this.pointsToSpend = buffer.readInt();
     }
 
     public void encode(FriendlyByteBuf buffer) {
@@ -47,6 +46,7 @@ public class ToolAttributePacket {
         buffer.writeFloat(this.value);
         buffer.writeInt(this.id);
         buffer.writeInt(this.slotIndex);
+        buffer.writeInt(this.pointsToSpend);
     }
 
     public void handle(Supplier<NetworkEvent.Context> ctx) {
@@ -54,6 +54,10 @@ public class ToolAttributePacket {
         ItemStack heldTool = getItemStack(playerEntity, this.slotIndex);
 
         if (!heldTool.isEmpty()) {
+            int skillPoints = (int) NBTUtils.getFloatOrAddKey(heldTool, "skill_points");
+            if (skillPoints == 0) return;
+            int pointsToAdd = Math.min(skillPoints, this.pointsToSpend);
+
             Enchantment enchantment = enchantmentFromString(this.key);
 
             if (enchantment != null) {
@@ -72,12 +76,12 @@ public class ToolAttributePacket {
                 // also add it like normal
                 NBTUtils.setValue(heldTool, this.key, this.value);
             } else {
-                NBTUtils.addValueToTag(heldTool, this.key, this.value);
+                NBTUtils.addValueToTag(heldTool, this.key, this.value * pointsToAdd);
             }
 
             // update the skill points array
             if (this.id != -1) {
-                NBTUtils.addValueToArray(heldTool, "points", this.id, 1);
+                NBTUtils.addValueToArray(heldTool, "points", this.id, pointsToAdd);
             }
         }
     }